refactor(blog): refactor blog pages code
