--- src/pages/blog/DateBook.tsx
@@ -1,17 +1,0 @@
-import React from 'react';
-import { Article } from 'service/nip/23';
-import { Book } from './Book';
-import { useDateBookData } from 'hooks/useDateBookData';
-
-export function DateBook({ articles }: { articles: Article[] }) {
-  const books = useDateBookData(articles);
-  return (
-    <span>
-      {books.map((book, key) => (
-        <span key={key} style={{ paddingRight: '20px' }}>
-          <Book title={book.title} count={book.count} time={book.time} />
-        </span>
-      ))}
-    </span>
-  );
-}
