+++ src/pages/blog/components/Collection/index.module.scss
@@ -1,0 +1,51 @@
+.collection {
+  h3{
+    margin-top: 30px;
+    margin-bottom: 15px;
+  }
+
+  .list{
+    display: flex;
+    width: 100%;
+    flex-flow: wrap;
+    justify-content: space-between;
+  }
+}
+
+.item{
+  display: flex;
+  flex-direction: column;
+  justify-content: space-between;
+  margin: 0 0 10px;
+  padding: 10px 0;
+  width: 150px;
+  cursor: pointer;
+  border: 1px solid #ccc;
+  border-radius: 8px;
+
+  &:hover {
+    border: 1px solid rgb(141, 197, 63);
+  }
+
+  .title{
+    margin: 0;
+    font-size: 18px;
+    font-weight: bold;
+    text-align: center;
+    text-transform: capitalize;
+  }
+
+  .count{
+    margin: 4px 0 0;
+    font-size: 14px;
+    text-align: center;
+    color: gray;
+  }
+
+  .time{
+    margin: 20px 0 0;
+    font-size: 12px;
+    text-align: center;
+    color: gray;
+  }
+}
